backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.49
alembic==1.12.1
psycopg2-binary==2.9.11
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# pydantic 2.11+ / asyncpg 0.30+ ship cp313 Windows wheels (no Rust build)
pydantic==2.11.7
email-validator==2.3.0
pydantic-settings==2.7.1
websockets==12.0
asyncpg==0.30.0
aioredis==2.0.1
python-dotenv==1.2.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
